# docs/implementation/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pongdang API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/data/water-index/assessments": {
      "get": {
        "tags": [
          "water-index"
        ],
        "summary": "Assessments",
        "operationId": "assessments_api_data_water_index_assessments_get",
        "parameters": [
          {
            "name": "spot_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Spot Id"
            }
          },
          {
            "name": "activity",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "swim",
                "surf",
                "relax",
                "mudflat",
                "onsen",
                "rafting"
              ],
              "type": "string",
              "title": "Activity"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "From"
            }
          },
          {
            "name": "until",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Until"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "name": "profile_id",
            "in": "query",
            "required": true,
            "schema": {
              "const": "general",
              "type": "string",
              "title": "Profile Id"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "observation",
                "forecast"
              ],
              "type": "string",
              "title": "Mode"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/data/water-forecast/forecasts": {
      "get": {
        "tags": [
          "water-forecast"
        ],
        "summary": "Forecasts",
        "operationId": "forecasts_api_data_water_forecast_forecasts_get",
        "parameters": [
          {
            "name": "spot_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Spot Id"
            }
          },
          {
            "name": "activity",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "swim",
                "surf",
                "relax",
                "mudflat",
                "onsen",
                "rafting"
              ],
              "type": "string",
              "title": "Activity"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "From"
            }
          },
          {
            "name": "until",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Until"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "name": "profile_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "const": "general",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Profile Id"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "observation",
                    "forecast"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssessmentDTO": {
        "properties": {
          "valid_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          },
          "contract_version": {
            "type": "string",
            "const": "water-assessment.v1-draft",
            "title": "Contract Version",
            "default": "water-assessment.v1-draft"
          },
          "target_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Target Id"
          },
          "assessment_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Id"
          },
          "input_manifest_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Manifest Id"
          },
          "spot_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Spot Id"
          },
          "activity": {
            "type": "string",
            "enum": [
              "swim",
              "surf",
              "relax",
              "mudflat",
              "onsen",
              "rafting"
            ],
            "title": "Activity"
          },
          "context": {
            "$ref": "#/components/schemas/Context"
          },
          "target": {
            "$ref": "#/components/schemas/Target"
          },
          "as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "evaluated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluated At"
          },
          "queried_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queried At"
          },
          "mode": {
            "type": "string",
            "enum": [
              "observation",
              "forecast",
              "mixed",
              "none"
            ],
            "title": "Mode",
            "default": "none"
          },
          "score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "availability": {
            "type": "string",
            "enum": [
              "available",
              "partial",
              "unavailable",
              "unsupported",
              "unknown"
            ],
            "title": "Availability",
            "default": "unknown"
          },
          "assessment_status": {
            "type": "string",
            "enum": [
              "evaluated",
              "not_evaluable",
              "withheld",
              "unsupported",
              "support_unknown",
              "outside_forecast_horizon",
              "model_unimplemented"
            ],
            "title": "Assessment Status"
          },
          "reason_codes": {
            "items": {
              "type": "string",
              "maxLength": 200,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 100,
            "title": "Reason Codes",
            "default": []
          },
          "safety_status": {
            "type": "string",
            "enum": [
              "restricted",
              "unknown",
              "caution",
              "no_known_restriction",
              "not_assessed"
            ],
            "title": "Safety Status"
          },
          "support": {
            "$ref": "#/components/schemas/SupportDTO"
          },
          "safety": {
            "$ref": "#/components/schemas/SafetyDTO"
          },
          "environment": {
            "$ref": "#/components/schemas/EnvironmentDTO"
          },
          "preference": {
            "$ref": "#/components/schemas/PreferenceDTO"
          },
          "recommendation": {
            "$ref": "#/components/schemas/RecommendationDTO"
          },
          "data_quality": {
            "$ref": "#/components/schemas/DataQuality"
          },
          "model": {
            "$ref": "#/components/schemas/ModelDTO"
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/InputDTO"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Inputs",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "target_id",
          "spot_id",
          "activity",
          "target",
          "assessment_status",
          "safety_status",
          "support",
          "safety",
          "environment",
          "recommendation",
          "model"
        ],
        "title": "AssessmentDTO"
      },
...
      "Context": {
        "properties": {
          "profile_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "equipment_profile_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipment Profile Id"
          },
          "skill_profile_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Profile Id"
          },
          "exposure_duration_minutes": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Exposure Duration Minutes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Context"
      },
      "Coverage": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "available",
              "partial",
              "unavailable",
              "unknown"
            ],
            "title": "Status"
          },
          "supported_windows": {
            "items": {
              "$ref": "#/components/schemas/TimeSpan"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Supported Windows"
          },
          "uncovered_windows": {
            "items": {
              "$ref": "#/components/schemas/TimeSpan"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Uncovered Windows"
          },
          "missing_targets": {
            "items": {
              "$ref": "#/components/schemas/MissingTarget"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Missing Targets"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Reason Codes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "supported_windows",
          "uncovered_windows",
          "missing_targets",
          "reason_codes"
        ],
        "title": "Coverage"
      },
...
      "Envelope": {
        "properties": {
          "contract_version": {
            "type": "string",
            "const": "water-assessment.v1-draft",
            "title": "Contract Version",
            "default": "water-assessment.v1-draft"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "queried_at": {
            "type": "string",
            "format": "date-time",
            "title": "Queried At"
          },
          "query": {
            "$ref": "#/components/schemas/ReadQuery"
          },
          "coverage": {
            "$ref": "#/components/schemas/Coverage"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/AssessmentDTO"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Rows"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "as_of",
          "queried_at",
          "query",
          "coverage",
          "rows",
          "total",
          "page",
          "page_size"
        ],
        "title": "Envelope"
      },
...
      "ForecastEnvelope": {
        "properties": {
          "contract_version": {
            "type": "string",
            "const": "water-forecast.v1",
            "title": "Contract Version",
            "default": "water-forecast.v1"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "queried_at": {
            "type": "string",
            "format": "date-time",
            "title": "Queried At"
          },
          "page": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ForecastView"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Rows"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "no_forecast_data",
              "outside_forecast_horizon",
              "missing_within_horizon"
            ],
            "title": "Status"
          },
          "horizon_start_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Horizon Start At"
          },
          "horizon_end_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Horizon End At"
          },
          "range_semantics": {
            "type": "string",
            "title": "Range Semantics"
          },
          "assessment_endpoint": {
            "type": "string",
            "const": "/api/data/water-index/assessments",
            "title": "Assessment Endpoint"
          },
          "recommendation_status": {
            "type": "string",
            "const": "unavailable",
            "title": "Recommendation Status"
          },
          "recommendation_reason": {
            "type": "string",
            "const": "model_not_validated",
            "title": "Recommendation Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "as_of",
          "queried_at",
          "page",
          "page_size",
          "total",
          "rows",
          "status",
          "horizon_start_at",
          "horizon_end_at",
          "range_semantics",
          "assessment_endpoint",
          "recommendation_status",
          "recommendation_reason"
        ],
        "title": "ForecastEnvelope"
      },
...
      "ForecastView": {
        "properties": {
          "contract_version": {
            "type": "string",
            "const": "water-forecast.v1",
            "title": "Contract Version",
            "default": "water-forecast.v1"
          },
          "source_key": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Source Key"
          },
          "adapter_version": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Adapter Version",
            "default": "1"
          },
          "spot_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Spot Id"
          },
          "station_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Station Id"
          },
          "station_code": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Station Code"
          },
          "station_name": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Station Name"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "provider": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Provider"
          },
          "source_record_id": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Source Record Id"
          },
          "provider_record_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Provider Record Id"
          },
          "snapshot_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Snapshot Id"
          },
          "issued_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issued At"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          },
          "target_start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Target Start At"
          },
          "target_end_at": {
            "type": "string",
            "format": "date-time",
            "title": "Target End At"
          },
          "timezone": {
            "type": "string",
            "const": "Asia/Seoul",
            "title": "Timezone",
            "default": "Asia/Seoul"
          },
          "data_kind": {
            "type": "string",
            "const": "official_forecast",
            "title": "Data Kind",
            "default": "official_forecast"
          },
          "spatial_relation": {
            "type": "string",
            "enum": [
              "station_observation_point",
              "representative_station"
            ],
            "title": "Spatial Relation"
          },
          "spatial_scope": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Spatial Scope"
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/InputDTO"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Inputs"
          },
          "revision_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Revision Id"
          },
          "previous_revision_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Revision Id"
          },
          "available_at": {
            "type": "string",
            "format": "date-time",
            "title": "Available At"
          },
          "requested_spot_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Requested Spot Id"
          },
          "mapping_evidence_ref": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Mapping Evidence Ref"
          },
          "state": {
            "type": "string",
            "enum": [
              "stale",
              "partial",
              "available"
            ],
            "title": "State"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Reason Codes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_key",
          "spot_id",
          "station_id",
          "station_code",
          "station_name",
          "provider",
          "source_record_id",
          "provider_record_id",
          "snapshot_id",
          "issued_at",
          "fetched_at",
          "target_start_at",
          "target_end_at",
          "spatial_relation",
          "spatial_scope",
          "inputs",
          "revision_id",
          "available_at",
          "requested_spot_id",
          "state",
          "reason_codes"
        ],
        "title": "ForecastView"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MissingTarget": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "instant",
              "interval"
            ],
            "title": "Kind"
          },
          "start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Start At"
          },
          "end_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End At"
          },
          "timezone": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Timezone"
          },
          "target_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Target Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "start_at",
          "timezone",
          "target_id"
        ],
        "title": "MissingTarget"
      },
...
      "PreferenceDTO": {
        "properties": {
          "status": {
            "type": "string",
            "const": "not_modelled",
            "title": "Status",
            "default": "not_modelled"
          },
          "score": {
            "type": "null",
            "title": "Score"
          },
          "ranking": {
            "type": "null",
            "title": "Ranking"
          },
          "comparison_scope": {
            "type": "null",
            "title": "Comparison Scope"
          },
          "reason_codes": {
            "items": {
              "type": "string",
              "maxLength": 200,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 100,
            "title": "Reason Codes",
            "default": [
              "preference_not_modelled"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PreferenceDTO"
      },
...
      "ReadQuery": {
        "properties": {
          "spot_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Spot Id"
          },
          "activity": {
            "type": "string",
            "enum": [
              "swim",
              "surf",
              "relax",
              "mudflat",
              "onsen",
              "rafting"
            ],
            "title": "Activity"
          },
          "profile_id": {
            "type": "string",
            "const": "general",
            "title": "Profile Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "observation",
              "forecast"
            ],
            "title": "Mode"
          },
          "from": {
            "type": "string",
            "format": "date-time",
            "title": "From"
          },
          "until": {
            "type": "string",
            "format": "date-time",
            "title": "Until"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "spot_id",
          "activity",
          "profile_id",
          "mode",
          "from",
          "until"
        ],
        "title": "ReadQuery"
      },
...
      "Target": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "instant",
              "interval"
            ],
            "title": "Kind"
          },
          "start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Start At"
          },
          "end_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End At"
          },
          "timezone": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Timezone"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "start_at",
          "timezone"
        ],
        "title": "Target"
      },
...
      "TimeSpan": {
        "properties": {
          "start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Start At"
          },
          "end_at": {
            "type": "string",
            "format": "date-time",
            "title": "End At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "start_at",
          "end_at"
        ],
        "title": "TimeSpan"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}